backend/requirements.txt
alembic
pydantic
SQLAlchemy
python-dotenv
fastapi
psycopg2-binary
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart